Lib Refactor
